Thanking all the devs
I don't know  how to do a REAL thanks to them. I mean, i'm using their work to put my shit on it, and that's bad i think. But the thing is: I have to thank them for all the work who made day and night, with all the headaches, with all the time spend on it, to show us why android is what it is: A great software to be in, to work and to use it daily. To bring the fun with new stuff, to make it our style freely. And more...